crypto_openssl: fix less-than-zero comparison of an unsigned value
Error from cppcheck:
src/crypto_openssl.c:696:21: style: Checking if unsigned expression 'pkcs7_out_len' is less than zero. [unsignedLessThanZero]
  if (pkcs7_out_len <= 0)

pkcs7_out_len was defined as a size_t. BIO_get_mem_data() returns a long.
Changed the type of pkcs7_out_len to long to clear the problem.

+++ src/crypto_openssl.c
@@ -546,7 +546,7 @@ pkcs7_generate_w_signature (unsigned char **pkcs7, size_t *pkcs7_size,
   EVP_PKEY *evp_pkey = NULL;
   const EVP_MD *evp_md = NULL;
   crypto_x509_t *x509 = NULL;
-  size_t pkcs7_out_len;
+  long pkcs7_out_len;
   unsigned char *key = NULL, *keyTmp, *crt = NULL, *out_bio_der = NULL;
   char *unnecessary_hdr = NULL, *unnecessary_name = NULL;
   long int keySize, crtSize;
